// src/components/business/profiles/profiles.types.ts
import type { ComputedRef, Ref } from 'vue'

export interface DirectoryOption {
    label: string
    value: string
}

export interface ProfilesDirectoryFilters {
    ageRange: string
    city: string
    heightRange: string
    education: string
    intentCode: string

    industry: string
    occupation: string
    language: string
    verified: string
    maritalStatus: string
    hasChildren: string
    acceptLongDistance: string
}

export type ProfilesSortKey =
    | 'recentActive'
    | 'verifiedFirst'
    | 'ageAsc'
    | 'ageDesc'

export interface UseProfilesDirectoryResult<TProfile> {
    filters: Ref<ProfilesDirectoryFilters>
    sortKey: Ref<ProfilesSortKey>
    page: Ref<number>
    pageSize: Ref<number>

    total: ComputedRef<number>
    totalPages: ComputedRef<number>
    filteredItems: ComputedRef<TProfile[]>
    sortedItems: ComputedRef<TProfile[]>
    pagedItems: ComputedRef<TProfile[]>
    pageStart: ComputedRef<number>
    pageEnd: ComputedRef<number>

    ageOptions: ComputedRef<DirectoryOption[]>
    cityOptions: ComputedRef<DirectoryOption[]>
    heightOptions: ComputedRef<DirectoryOption[]>
    educationOptions: ComputedRef<DirectoryOption[]>
    intentOptions: ComputedRef<DirectoryOption[]>
    industryOptions: ComputedRef<DirectoryOption[]>
    occupationOptions: ComputedRef<DirectoryOption[]>
    languageOptions: ComputedRef<DirectoryOption[]>
    verifiedOptions: ComputedRef<DirectoryOption[]>
    maritalStatusOptions: ComputedRef<DirectoryOption[]>
    childrenOptions: ComputedRef<DirectoryOption[]>
    longDistanceOptions: ComputedRef<DirectoryOption[]>
    sortOptions: ComputedRef<DirectoryOption[]>

    updateFilters: (nextFilters: Partial<ProfilesDirectoryFilters>) => void
    resetFilters: () => void
    updateSort: (nextSortKey: string) => void
    changePage: (nextPage: number) => void
}